unify type names, move gql queries closer to components that use them
