+++ admin-ui/src/common/queries.ts
@@ -300,33 +300,6 @@ export const DELETE_RESOURCE = gql`
   }
 `;
 
-export const RESERVATIONS_QUERY = gql`
-  query reservations($state: String) {
-    reservations(state: $state) {
-      edges {
-        node {
-          pk
-          workingMemo
-          state
-          reservationUnits {
-            nameFi
-            unit {
-              nameFi
-            }
-          }
-          begin
-          end
-          reserveeFirstName
-          reserveeLastName
-          reserveeEmail
-          name
-          price
-        }
-      }
-    }
-  }
-`;
-
 export const HANDLING_COUNT_QUERY = gql`
   query handlingCount {
     reservations(state: "REQUIRES_HANDLING") {
@@ -334,115 +307,6 @@ export const HANDLING_COUNT_QUERY = gql`
         node {
           pk
         }
-      }
-    }
-  }
-`;
-
-export const RESERVATION_QUERY = gql`
-  query reservationByPk($pk: Int!) {
-    reservationByPk(pk: $pk) {
-      pk
-      workingMemo
-      reservationUnits {
-        nameFi
-        unit {
-          nameFi
-        }
-      }
-      ageGroup {
-        minimum
-        maximum
-      }
-      purpose {
-        nameFi
-      }
-      numPersons
-      name
-      price
-      unitPrice
-      description
-      reserveeFirstName
-      reserveeLastName
-      reserveePhone
-      begin
-      end
-      calendarUrl
-      user
-      state
-      reserveeOrganisationName
-      reserveeEmail
-      reserveeId
-      reserveeIsUnregisteredAssociation
-      reserveeAddressStreet
-      reserveeAddressCity
-      reserveeAddressZip
-      billingFirstName
-      billingLastName
-      billingPhone
-      billingEmail
-      billingAddressStreet
-      billingAddressCity
-      billingAddressZip
-      freeOfChargeReason
-      applyingForFreeOfCharge
-    }
-  }
-`;
-
-export const APPROVE_RESERVATION = gql`
-  mutation approveReservation($input: ReservationApproveMutationInput!) {
-    approveReservation(input: $input) {
-      errors {
-        field
-        messages
-      }
-    }
-  }
-`;
-
-export const DENY_RESERVATION = gql`
-  mutation denyReservation($input: ReservationDenyMutationInput!) {
-    denyReservation(input: $input) {
-      errors {
-        field
-        messages
-      }
-    }
-  }
-`;
-
-export const REQUIRE_HANDLING_RESERVATION = gql`
-  mutation requireHandling($input: ReservationRequiresHandlingMutationInput!) {
-    requireHandlingForReservation(input: $input) {
-      errors {
-        field
-        messages
-      }
-    }
-  }
-`;
-
-export const RESERVATION_DENY_REASONS = gql`
-  query reservationDenyReasons {
-    reservationDenyReasons {
-      edges {
-        node {
-          pk
-          reasonFi
-        }
-      }
-    }
-  }
-`;
-
-export const UPDATE_WORKING_MEMO = gql`
-  mutation updateWorkingMemo($input: ReservationWorkingMemoMutationInput!) {
-    updateReservationWorkingMemo(input: $input) {
-      workingMemo
-      errors {
-        field
-        messages
       }
     }
   }
