+++ admin-ui/src/component/reservations/requested/queries.ts
@@ -1,0 +1,139 @@
+import { gql } from "@apollo/client";
+
+export const UPDATE_WORKING_MEMO = gql`
+  mutation updateWorkingMemo($input: ReservationWorkingMemoMutationInput!) {
+    updateReservationWorkingMemo(input: $input) {
+      workingMemo
+      errors {
+        field
+        messages
+      }
+    }
+  }
+`;
+
+export const RESERVATION_QUERY = gql`
+  query reservationByPk($pk: Int!) {
+    reservationByPk(pk: $pk) {
+      pk
+      workingMemo
+      reservationUnits {
+        nameFi
+        unit {
+          nameFi
+        }
+      }
+      ageGroup {
+        minimum
+        maximum
+      }
+      purpose {
+        nameFi
+      }
+      numPersons
+      reserveeType
+      reserveeIsUnregisteredAssociation
+      name
+      price
+      unitPrice
+      description
+      reserveeFirstName
+      reserveeLastName
+      reserveePhone
+      begin
+      end
+      calendarUrl
+      user
+      state
+      reserveeOrganisationName
+      reserveeEmail
+      reserveeId
+      reserveeIsUnregisteredAssociation
+      reserveeAddressStreet
+      reserveeAddressCity
+      reserveeAddressZip
+      billingFirstName
+      billingLastName
+      billingPhone
+      billingEmail
+      billingAddressStreet
+      billingAddressCity
+      billingAddressZip
+      freeOfChargeReason
+      applyingForFreeOfCharge
+    }
+  }
+`;
+
+export const APPROVE_RESERVATION = gql`
+  mutation approveReservation($input: ReservationApproveMutationInput!) {
+    approveReservation(input: $input) {
+      errors {
+        field
+        messages
+      }
+    }
+  }
+`;
+
+export const DENY_RESERVATION = gql`
+  mutation denyReservation($input: ReservationDenyMutationInput!) {
+    denyReservation(input: $input) {
+      errors {
+        field
+        messages
+      }
+    }
+  }
+`;
+
+export const REQUIRE_HANDLING_RESERVATION = gql`
+  mutation requireHandling($input: ReservationRequiresHandlingMutationInput!) {
+    requireHandlingForReservation(input: $input) {
+      errors {
+        field
+        messages
+      }
+    }
+  }
+`;
+
+export const RESERVATION_DENY_REASONS = gql`
+  query reservationDenyReasons {
+    reservationDenyReasons {
+      edges {
+        node {
+          pk
+          reasonFi
+        }
+      }
+    }
+  }
+`;
+
+export const RESERVATIONS_QUERY = gql`
+  query reservations($state: String) {
+    reservations(state: $state) {
+      edges {
+        node {
+          pk
+          workingMemo
+          state
+          reservationUnits {
+            nameFi
+            unit {
+              nameFi
+            }
+          }
+          begin
+          end
+          reserveeFirstName
+          reserveeLastName
+          reserveeEmail
+          name
+          price
+        }
+      }
+    }
+  }
+`;
